Renaming data table that is in microservice.tsx
